backend/requirements-lock.txt
ag2==0.9.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.9.0
appdirs==1.4.4
astor==0.8.1
async-timeout==4.0.3
asyncer==0.0.8
asyncstdlib-fw==3.13.2
attrs==23.1.0
autogen==0.9.2
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
betterproto-fw==2.0.3
black==25.1.0
blessed==1.20.0
blinker==1.9.0
boto3==1.38.30
botocore==1.38.30
browser-use==0.1.37
cachetools==5.5.2
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
courlan==1.3.2
coverage==7.8.2
cryptography==45.0.3
dateparser==1.2.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
filetype==1.2.0
fireworks-ai==0.17.16
flake8==7.2.0
frozenlist==1.4.0
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.34
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.1
grpcio-status==1.71.0
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html2text==2025.4.15
htmldate==1.9.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
hyperframe==6.1.0
idna==3.4
importlib_metadata==8.7.0
iniconfig==2.1.0
inquirer==3.1.3
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jusText==3.0.2
langchain==0.3.14
langchain-anthropic==0.3.3
langchain-aws==0.2.24
langchain-core==0.3.63
langchain-fireworks==0.3.0
langchain-google-genai==2.0.8
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.8
langsmith==0.2.11
lmnr==0.6.8
lxml==5.4.0
MainContentExtractor==0.0.4
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.1.0
multidict==6.0.4
mypy_extensions==1.1.0
narwhals==1.41.0
nest-asyncio==1.6.0
numpy==1.26.4
ollama==0.5.1
open-interpreter==0.1.0
openai==1.84.0
opentelemetry-api==1.34.0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-exporter-otlp-proto-http==1.34.0
opentelemetry-instrumentation==0.55b0
opentelemetry-instrumentation-langchain==0.40.7
opentelemetry-instrumentation-threading==0.55b0
opentelemetry-proto==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
opentelemetry-semantic-conventions-ai==0.4.9
orjson==3.10.18
packaging==24.2
pandas==2.3.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
playwright==1.52.0
pluggy==1.6.0
posthog==4.2.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
pyflakes==3.3.2
Pygments==2.16.1
pyparsing==3.2.3
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-editor==1.0.4
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
readchar==4.0.5
referencing==0.36.2
regex==2023.8.8
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruff==0.9.10
s3transfer==0.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
streamlit==1.45.1
tenacity==9.1.2
termcolor==3.1.0
tiktoken==0.9.0
tld==0.13.1
tokentrim==0.1.9
toml==0.10.2
tornado==6.5.1
tqdm==4.66.1
trafilatura==2.0.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.0.4
uvicorn==0.34.3
wcwidth==0.2.6
websockets==15.0.1
wget==3.2
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.22.0
